% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{COAD.tumorMarkerParams.hg19}
\alias{COAD.tumorMarkerParams.hg19}
\title{Parameters of COAD markers}
\value{
Downloads and caches a `.txt.gz` file.
}
\description{
The shape parameters of beta distribution characterizing methylation markers 
associated with colon adenocarcinoma (COAD)
}
\examples{

COADMarkerParams <- COAD.tumorMarkerParams.hg19()
COADMarkerParams

}
