% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{LIHC.tumorMarkerParams.hg19}
\alias{LIHC.tumorMarkerParams.hg19}
\title{Parameters of LIHC markers}
\value{
Downloads and caches a `.txt.gz` file.
}
\description{
The shape parameters of beta distribution characterizing methylation markers 
associated with liver hepatocellular carcinoma (LIHC)
}
\examples{

LIHCMarkerParams <- LIHC.tumorMarkerParams.hg19()
LIHCMarkerParams

}
