% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{LUNG.tumorMarkerParams.hg19}
\alias{LUNG.tumorMarkerParams.hg19}
\title{Parameters of lung cancer markers}
\value{
Downloads and caches a `.txt.gz` file.
}
\description{
The shape parameters of beta distribution characterizing methylation markers 
associated with lung cancer
}
\examples{

LUNGMarkerParams <- LUNG.tumorMarkerParams.hg19()
LUNGMarkerParams

}
