% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-tissueMarkerParams_annot.R
\name{tissueMarkerParams.annot}
\alias{tissueMarkerParams.annot}
\title{Annotation of markers in the tissueMarkerParams.hg19}
\format{
A data frame containing the tissue-specific marker positions aligned
to the HG19 reference genome, the type of strategy used to identify them,
and the tissue groups for pair comparison.
\describe{
\item{chrom}{Chromosome}
\item{start}{Chromosome start}
\item{end}{Chromosome end}
\item{marker_type}{The type of strategy used to identify markers}
\item{pair_comparison}{The tissue groups for pair comparison}

}
}
\usage{
data("tissueMarkerParams.annot")
}
\value{
A data frame with 51035 rows and 5 variables
}
\description{
Label tissue markers with the strategies used to identify them. Details
can be found in paper https://doi.org/10.1073/pnas.2305236120
}
\author{
Ran Hu \email{huran@ucla.edu}
}
