% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{tissueMarkerParams.hg19}
\alias{tissueMarkerParams.hg19}
\title{Parameters of tissue markers}
\value{
Downloads and caches a `.txt.gz` file.
}
\description{
The shape parameters of beta distribution characterizing methylation markers 
specific to 29 primary human tissue types
}
\examples{

tissueMarkerParams <- tissueMarkerParams.hg19()
tissueMarkerParams

}
