% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-LUNG.tumorMarkerParams.R
\name{LUNG.tumorMarkerParams.hg19}
\alias{LUNG.tumorMarkerParams.hg19}
\title{Methylation patterns of lung cancer markers}
\format{
A data frame containing the lung cancer-specific marker positions
aligned to the HG19 reference genome and the beta distribution shape
parameters of the markers.
\describe{
\item{chrom}{Chromosome}
\item{start}{Chromosome start}
\item{end}{Chromosome end}
\item{LUNG_normal}{Marker parameters of normal lung tissue}
\item{LUNG_tumor}{Marker parameters of lung tumor tissue}
\item{WBC}{Marker parameters of white blood cell}
\item{normal_plasma}{Marker parameters of normal plasma}
\item{type}{Marker type}
}
}
\usage{
data("LUNG.tumorMarkerParams.hg19")
}
\value{
A data frame with 20000 rows and 8 variables
}
\description{
Methylation patterns of markers associated with lung cancer
}
\author{
Ran Hu \email{huran@ucla.edu}
}
