% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{STAD.tumorMarkerParams.hg19}
\alias{STAD.tumorMarkerParams.hg19}
\title{Parameters of STAD markers}
\value{
Downloads and caches a `.txt.gz` file.
}
\description{
The shape parameters of beta distribution characterizing methylation markers 
associated with stomach adenocarcinoma (STAD)
}
\examples{

STADMarkerParams <- STAD.tumorMarkerParams.hg19()
STADMarkerParams

}
