% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{tissueMarkerParams.annot}
\alias{tissueMarkerParams.annot}
\title{Annotation of tissue markers}
\value{
Downloads and caches a `.txt.gz` file.
}
\description{
The annotation of strategies used to identify tissue markers.
}
\examples{

tissueMarkerAnnot <- tissueMarkerParams.annot()
tissueMarkerAnnot

}
